requirements.txt
fastapi>=0.68.0
uvicorn>=0.15.0
neo4j>=5.0.0
pydantic>=1.8.0
google-generativeai>=0.3.0
python-jose>=3.3.0
passlib>=1.7.4
python-multipart>=0.0.5
bcrypt>=3.2.0
streamlit>=1.28.0
requests>=2.31.0
python-dotenv>=1.0.0